use poll_event instead of peek, and wrap in if() to allow idling between inputs

+++ term.h
@@ -45,7 +45,7 @@ void checkInput( struct tb_event *ev, Color *color, const uint16_t **c, const bl
 void setColor( Color *restrict color, uint32_t *restrict c );
 
 // draw a horizontal line
-void hLine( uint16_t x, uint16_t y, uint32_t fgCol, uint32_t bgCol, uint16_t c, uint8_t dir );
+void hLine( uint8_t x, uint16_t y, uint32_t fgCol, uint32_t bgCol, uint16_t c, uint8_t dir );
 
 // draw a vertical line
 void vLine( uint16_t x, uint16_t y, uint32_t fgCol, uint32_t bgCol, uint16_t c, uint8_t dir );
